Improve `LUA_DEBUG`

+++ sdk/include/luabind/class.hpp
@@ -718,7 +718,7 @@ namespace luabind
 #else
 			void add_setter(
 				const char* name
-				, std::function<int(lua_State*, int)> s
+				, std::function<int(lua_State*, ptrdiff_t)> s
 				, int (*match)(lua_State*, int)
 				, void (*get_sig_ptr)(lua_State*, string_class&));
 #endif
